Use leftLabel and default styling in suggestions.
It's uncommon for autocomplete providers to use the description box for return types, so I have instead used the leftLabel property to display the return type. This is a right-aligned column placed before the displayText that, by default, is slightly dimmed and is easier to read.

In addition, the description field is now used to show required namespace imports, and the type will use the default colorization in autocomplete for a few types. This could be expanded to correctly map each kind in C# to the default types in autocomplete, but for now it's able to colorize properties and methods similarly to other completion providers, in addition to the icons used in OmniSharp.